backend/requirements.txt
fyers-apiv3
Flask
python-dotenv
